backend/pyproject.toml
[project]
name = "bidding-flash-sale-system-backend"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.123.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "alembic>=1.13.0",
    "pre-commit>=4.5.0",
    "bcrypt==4.1.3",
]
